Move moveable into a separate library.
